+++ resources/lang/en/account.php
@@ -2,6 +2,11 @@
 
 return [
 
-    'register_success' => 'Register success!',
+    'register_success'  => 'Register success!',
+    'order_list'        => 'Order list',
+    'my_profile'        => 'My profile',
+    'change_password'   => 'Change password',
+    'change_infomation' => 'Change infomation',
+    'update_infomation' => 'Update infomation',
 
 ];