+++ resources/lang/vi/account.php
@@ -2,6 +2,10 @@
 
 return [
 
-    'register_success' => 'Đăng ký tài khoàn thành công!',
-
+    'register_success'  => 'Đăng ký tài khoàn thành công!',
+    'order_list'        => 'Danh sách đơn hàng',
+    'my_profile'        => 'Tài khoản của tôi',
+    'change_password'   => 'Thay đổi mật khẩu',
+    'change_infomation' => 'Thay đổi thông tin',
+    'update_infomation' => 'Cập nhật thông tin',
 ];